Changed the OAuth package to specify which version of OAuth it supports.
