+++ src/main/java/com/synopsys/integration/jira/common/rest/oauth1a/JiraOAuthGetAccessToken.java
@@ -20,13 +20,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.jira.common.rest.oauth;
+package com.synopsys.integration.jira.common.rest.oauth1a;
 
-import com.google.api.client.auth.oauth.OAuthGetTemporaryToken;
+import com.google.api.client.auth.oauth.OAuthGetAccessToken;
 
-public class JiraOAuthGetTemporaryToken extends OAuthGetTemporaryToken {
-    public JiraOAuthGetTemporaryToken(String temporaryTokenRequestUrl) {
-        super(temporaryTokenRequestUrl);
+public class JiraOAuthGetAccessToken extends OAuthGetAccessToken {
+    public JiraOAuthGetAccessToken(String accessTokenRequestUrl) {
+        super(accessTokenRequestUrl);
         this.usePost = true;
     }
